+++ airbyte-cdk/java/airbyte-cdk/s3-destinations/src/main/java/io/airbyte/cdk/integrations/destination/s3/parquet/S3ParquetFormatConfig.java
@@ -50,7 +50,7 @@ public class S3ParquetFormatConfig implements S3FormatConfig {
     return compressionCodec;
   }
 
-  public int getBlockSize() {
+  public long getBlockSize() {
     return blockSize;
   }
